+++ lib/flutter_state_widget.dart
@@ -1,4 +1,5 @@
 export 'src/load_state.dart';
 export 'src/state_snapshot.dart';
 export 'src/state_widget.dart';
+export 'src/state_widget_theme.dart';
 export 'src/state_widget_texts.dart';